refactor(connector): own Clipper CLI and MCP access
