--- packages/cli/src/utils/local-bridge.ts
@@ -1,2 +1,0 @@
-// Backward-compatible source entry point. Local Bridge is owned by Connector.
-export * from '@shadowob/connector/local-bridge'
